// src/shared/types.ts
export type SanityImageProps = ImageQueryInputs & {
  preview?: string;

  /**
   * The base url for the Sanity CDN including project ID and dataset. If not
   * provided, the `projectId` and `dataset` props will be used to construct the
   * URL.
   */
  baseUrl?: string;

  /**
   * The Sanity project ID to use for the image URL. If preferred, the `baseUrl`
   * prop can be provided instead.
   */
  projectId?: string;

  /**
   * The Sanity dataset to use for the image URL. If preferred, the `baseUrl`
   * prop can be provided instead.
   */
  dataset?: string;

  /**
   * `alt` attribute for the image; set to an empty string if not provided.
   */
  // alt?: string

  /**
   * Passed through to the <img> tag as `height`, overriding the `aspectRatio`
   * option, if enabled.
   */
  htmlHeight?: number;

  /**
   * Passed through to the <img> tag as `width`, overriding the `aspectRatio`
   * option, if enabled.
   */
  htmlWidth?: number;

  /**
   * Passed through to the <img> tag as `id` since the `id` prop is reserved for
   * the Sanity image asset ID.
   */
  htmlId?: string;

  /**
   * Query string params to pass to Sanity's image CDN directly. Note that this
   * is only a subset of the params supported by the Sanity image CDN. Many are
   * set automatically by this library, and several others result in behavior
   * you probably don't want. If you need something and have a compelling use
   * case, please open an issue and I'd be delighted to consider it.
   */
  queryParams?: DirectQueryParams;
};

export type ImageWithPreviewProps<T extends React.ElementType> = {
  preview: string;
} & AsProp<T> &
  React.ComponentPropsWithRef<T>;

export type CropData = {
  bottom: number;
  left: number;
  right: number;
  top: number;
};

export type HotspotData = {
  x: number;
  y: number;
};

export type Asset = {
  _id: string;
  crop?: CropData;
  hotspot?: HotspotData;
};

export type ImageIdParts = {
  assetId: string;
  dimensions: {
    height: number;
    width: number;
    aspectRatio: number;
  };
  format: string;
};

/**
 * Query string params to pass to Sanity's image CDN directly.
 */
export type DirectQueryParams = {
  /**
   * Blur 1-2000.
   */
  blur?: number;

  /**
   * Flipping. Flip image horizontally, vertically or both. Possible values: h,
   * v, hv.
   */
  flip?: 'h' | 'v' | 'hv';

  /**
   * Format. Convert image to jpg, pjpg, png, or webp. Note: like other query
   * string params, this doesn't work on SVG source images.
   */
  fm?: 'jpg' | 'pjpg' | 'png' | 'webp';

  /**
   * Quality 0-100. Specify the compression quality (where applicable). Defaults
   * to 75 for JPG and WebP.
   */
  q?: number;

  /**
   * Saturation. Currently the asset pipeline only supports `sat=-100`, which
   * renders the image with grayscale colors. Support for more levels of
   * saturation is planned for later.
   */
  sat?: -100;

  /**
   * Sharpen 0-100.
   */
  sharpen?: number;
};

export type ImageQueryInputs = {
  /** The Sanity Image ID (`_id` or `_ref` field value) */
  id: string;

  /**
   * Use `cover` to crop the image to match the requested aspect ratio (based on
   * `width` and `height`). Use `contain` to fit the image to the boundaries
   * provided without altering the aspect ratio.
   * @default "contain"
   */
  mode?: 'cover' | 'contain';

  /**
   * The target width of the image in pixels. Only used for determining the
   * dimensions of the generated assets, not for layout. Use CSS to specify how
   * the browser should render the image instead.
   */
  width?: number;

  /**
   * The target height of the image in pixels. Only used for determining the
   * dimensions of the generated assets, not for layout. Use CSS to specify how
   * the browser should render the image instead.
   */
  height?: number;

  /**
   * The hotspot coordinates to use for the image. Note: hotspot `width` and
   * `height` are not used.
   */
  hotspot?: { x: number; y: number };

  /** The crop coordinates to use for the image. */
  crop?: CropData;

  queryParams?: DirectQueryParams;
};